Update tasks 0, 5, 6, 9, 100 and adding readme

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	int nodes;
+	size_t nodes;
 
 	nodes = 0;
 	while (h)
